Sort My Toys Implemented

+++ src/Pages/MyToys/MyToys.jsx
@@ -5,8 +5,13 @@ import { Link } from "react-router-dom";
 const MyToys = () => {
   const { user } = useContext(AuthContext);
   const [toyData, setToyData] = useState([]);
+  const [sort, setSort] = useState(true);
   const url = `http://localhost:5000/toydata?email=${user?.email}`;
-
+  useEffect(() => {
+    sort
+      ? toyData.sort((a, b) => a.price - b.price)
+      : toyData.sort((a, b) => b.price - a.price);
+  }, [sort, toyData]);
   useEffect(() => {
     fetch(url, {
       method: "GET",
@@ -20,21 +25,29 @@ const MyToys = () => {
         setToyData(d);
       });
   }, [url]);
+
   const handleDelete = (_id) => {
-    console.log(_id);
-    fetch(`http://localhost:5000/singletoy/${_id}`, {
-      method: "DELETE",
-    })
-      .then((r) => r.json())
-      .then((d) => {
-        console.log(d);
-        if (d.deletedCount > 0) {
-          alert("Successfully Removed!");
-        }
-      });
+    const deleteConfirm = confirm("Are you sure you want to remove it?");
+    if (deleteConfirm) {
+      fetch(`http://localhost:5000/singletoy/${_id}`, {
+        method: "DELETE",
+      })
+        .then((r) => r.json())
+        .then((d) => {
+          console.log(d);
+          if (d.deletedCount > 0) {
+            alert("Successfully Removed!");
+            const remainingToys = toyData.filter((toy) => toy._id !== _id);
+            setToyData(remainingToys);
+          }
+        });
+    }
   };
   return (
     <>
+      <button onClick={() => setSort(!sort)} className="btn my-4">
+        {sort ? "Sort Ascending" : "Sort Descending"}
+      </button>
       <div className="overflow-x-auto">
         <table className="table table-compact w-full">
           <thead>
@@ -56,21 +69,24 @@ const MyToys = () => {
             {toyData.map((toy, i) => (
               <>
                 <tr key={toy._id}>
-                  <th>{i}</th>
+                  <th>{i + 1}</th>
                   <td>
                     <img
                       style={{ width: "40px", height: "40px" }}
                       src={toy.toyPictureURL}
                     />
                   </td>
-                  <td>{toy.toyName}</td>
-                  <td>{toy.toySubCategory}</td>
-                  <td>{toy.sellerName}</td>
-                  <td>{toy.sellerEmail}</td>
-                  <td>{toy.price}</td>
-                  <td>{toy.rating}</td>
-                  <td>{toy.quantity}</td>
-                  <td>{toy.description?.slice(0, 8)}</td>
+                  <td>{toy?.toyName}</td>
+                  <td>{toy?.toySubCategory}</td>
+                  <td>{toy?.sellerName}</td>
+                  <td>{toy?.sellerEmail}</td>
+                  <td>{toy?.price}</td>
+                  <td>{toy?.rating}</td>
+                  <td>{toy?.quantity}</td>
+                  <td>
+                    {toy?.toyDescription?.slice(0, 10)}
+                    {"..."}
+                  </td>
 
                   <td className="flex justify-evenly">
                     <Link to={`/updatetoy/${toy._id}`}>
